requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
aiohttp=3.9.3=py312h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=14.0.2=h374c478_1
attrs=23.1.0=py312h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.10.55=h721c034_0
beautifulsoup4=4.12.2=py312h06a4308_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.7=py312ha883a20_0
brotli-python=1.0.9=py312h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
coverage=7.2.2=py312h5eee18b_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
datasets=2.18.0=pyhd8ed1ab_0
dill=0.3.7=py312h06a4308_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py312h5eee18b_0
fsspec=2023.10.0=py312h06a4308_0
gdown=5.1.0=pyhd8ed1ab_0
gflags=2.2.2=h6a678d5_1
glog=0.5.0=h6a678d5_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
huggingface_hub=0.21.4=pyhd8ed1ab_0
icu=73.1=h6a678d5_0
idna=3.4=py312h06a4308_0
ignite=0.4.13=py_0
importlib-metadata=7.0.1=py312h06a4308_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46306
isodate=0.6.1=py312h06a4308_0
jinja2=3.1.3=py312h06a4308_0
joblib=1.2.0=py312h06a4308_0
jpeg=9e=h5eee18b_1
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsqlite=3.45.2=h2797004_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py312h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
multidict=6.0.4=py312h5eee18b_0
multiprocess=0.70.15=py312h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py312h06a4308_0
nlpaug=1.1.11=pyhd8ed1ab_1
numexpr=2.8.7=py312hf827012_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_0
orc=1.7.4=hb3bc3d3_1
packaging=23.1=py312h06a4308_0
pandas=2.2.1=py312h526ad5a_0
pillow=10.2.0=py312h5eee18b_0
pip=23.3.1=py312h06a4308_0
pluggy=1.0.0=py312h06a4308_1
pyarrow=14.0.2=py312hb107042_0
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pyparsing=3.0.9=py312h06a4308_0
pysocks=1.7.1=py312h06a4308_0
pytest=7.4.0=py312h06a4308_0
python=3.12.2=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py312h5eee18b_1
python_abi=3.12=4_cp312
pytorch=2.2.1=py3.12_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py312h06a4308_0
pyyaml=6.0.1=py312h5eee18b_0
rdflib=7.0.0=pyhd8ed1ab_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py312h5eee18b_0
requests=2.31.0=py312h06a4308_1
s2n=1.3.27=hdbd6064_0
safetensors=0.4.2=py312h4b3b743_0
scikit-learn=1.3.0=py312h526ad5a_2
scipy=1.11.4=py312hc5e2394_0
sentencepiece=0.1.99=py312hdb19cb5_0
setuptools=68.2.2=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
soupsieve=2.5=py312h06a4308_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.15.2=py312hfef1a59_0
torchaudio=2.2.1=py312_cu121
torchvision=0.17.1=py312_cu121
tqdm=4.65.0=py312he106c6f_0
transformers=4.38.2=pyhd8ed1ab_0
typing-extensions=4.9.0=py312h06a4308_1
typing_extensions=4.9.0=py312h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py312h06a4308_1
utf8proc=2.6.1=h5eee18b_1
wheel=0.41.2=py312h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py312h5eee18b_0
zipp=3.17.0=py312h06a4308_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
